requirements.txt
python-dotenv
pandas
psycopg2-binary
neo4j
logging
altair
splink
pyspark